requirements.txt
python-telegram-bot==21.6
langchain==0.3.7
langchain-google-genai==2.0.5
google-generativeai==0.8.3
supabase==2.7.4
streamlit==1.40.0
pydantic==2.9.2
pandas==2.2.3
python-dotenv==1.0.1

